rename field and make sure to get into the database

+++ server/sql_utils.go
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS Exercise (
 	ID int NOT NULL AUTO_INCREMENT,
 	name varchar(255) NOT NULL,
     comment text,
-	fbID int NOT NULL,
+	fbID bigint NOT NULL,
     PRIMARY KEY(ID)
 );
